// Front End/src/app/models/user.model.ts
export class User {

  id: string;
  userName: string;
  password: string;
  firstName: string;
  lastName: string;
  contactNumber: string;
  regCode: string;
  role:string;
  linkedinUrl: string;
  yearsOfExperience: string;
  active:boolean;
  confirmedSignUp:boolean;
  resetPassword:boolean;
  resetPasswordDate:string;
  email: string;
}